refactoring day09
