+++ 2019/src/lib.rs
@@ -8,4 +8,5 @@ pub mod day05;
 pub mod day06;
 pub mod day07;
 pub mod day08;
+pub mod day09;
 pub mod day10;
